Gestion de la récupération du drapeau ennemi et sauvetage de drapeau de son équipe

+++ network/DatagramUpdateClient.java
@@ -11,8 +11,8 @@ import client.Model.Missile;
  */
 public class DatagramUpdateClient {
 	public int idPartie;
-	public double x;
-	public double y;
+	public float x;
+	public float y;
 	public float accelerationX;
 	public float accelerationY;
